+++ src/main/java/com/ricardo/traker/model/dto/request/AlertRequest/AlertRequestDto.java
@@ -19,7 +19,7 @@ public class AlertRequestDto {
 
     @JsonProperty("vehicle_id")
     @NotNull
-    private Integer vehicleId;
+    private Long vehicleId;
 
     private AlertArrivalRequestDto arrival;
