+++ src/main/java/com/ricardo/traker/model/dto/response/UserResponseDto.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @AllArgsConstructor
 public class UserResponseDto {
 
-    private Integer id;
+    private Long id;
 
     private String firstname;
