+++ src/main/java/com/ricardo/traker/repository/AlertArrivalRepository.java
@@ -3,5 +3,5 @@ package com.ricardo.traker.repository;
 import com.ricardo.traker.model.entity.AlertEntity.AlertArrivalEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AlertArrivalRepository extends JpaRepository<AlertArrivalEntity, Integer> {
+public interface AlertArrivalRepository extends JpaRepository<AlertArrivalEntity, Long> {
 }
