+++ src/main/java/com/ricardo/traker/repository/AlertDistanceRepository.java
@@ -3,5 +3,5 @@ package com.ricardo.traker.repository;
 import com.ricardo.traker.model.entity.AlertEntity.AlertDistanceEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AlertDistanceRepository extends JpaRepository<AlertDistanceEntity, Integer> {
+public interface AlertDistanceRepository extends JpaRepository<AlertDistanceEntity, Long> {
 }
