+++ src/main/java/com/ricardo/traker/repository/AlertSpeedRepository.java
@@ -3,5 +3,5 @@ package com.ricardo.traker.repository;
 import com.ricardo.traker.model.entity.AlertEntity.AlertSpeedEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AlertSpeedRepository extends JpaRepository<AlertSpeedEntity, Integer> {
+public interface AlertSpeedRepository extends JpaRepository<AlertSpeedEntity, Long> {
 }
